LayoutTests/imported/w3c/web-platform-tests/css/css-text/i18n/other-lang/css-text-line-break-de-in-loose-expected.txt


FAIL 2024  ONE DOT LEADER may appear at line start if de and loose assert_approx_equals: expected 48.697265625 +/- 1 but got 78.697265625
FAIL 2025  TWO DOT LEADER may appear at line start if de and loose assert_approx_equals: expected 59.009765625 +/- 1 but got 89.009765625
FAIL 2026  HORIZONTAL ELLIPSIS may appear at line start if de and loose assert_approx_equals: expected 69 +/- 1 but got 99
FAIL 22EF  MIDLINE HORIZONTAL ELLIPSIS may appear at line start if de and loose assert_approx_equals: expected 69 +/- 1 but got 99
FAIL FE19  PRESENTATION FORM FOR VERTICAL HORIZONTAL ELLIPSIS may appear at line start if de and loose assert_approx_equals: expected 69 +/- 1 but got 99
2024
文文文文文文․字字
文文文文文文
․字字
2025
文文文文文文‥字字
文文文文文文
‥字字
2026
文文文文文文…字字
文文文文文文
…字字
22EF
文文文文文文⋯字字
文文文文文文
⋯字字
FE19
文文文文文文︙字字
文文文文文文
︙字字
